+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog for Paco
 ==================
 
+7.1.1 (unreleased)
+------------------
+
+- Nothing changed yet.
+
+
 7.1.0 (2020-06-22)
 ------------------